Do not run FREEC post-processing when there are no CNVs
FREEC produces an empty _CNVs file when successful but copy number is
always 2. This is not a failure, but fails the other commands. Do not
run them in this situation, but still fail if the file is missing or the
ratio file is empty (the ratios are always produced). Continue producing
an empty PNG to keep checkers happy and for completeness.