Work-around bamUtil behaviour when diff is empty
bamUtil does not produce an empty diff file when the diff itself is
empty. This is annoying for checks - which see it as a failure - but
should not happen in normal cases. Thus, we leave this check to
downstream tools instead of failing on it here. The cases where it
happens warrant a manual inspection.

We could touch the diff file first and then overwrite it with the diff
to ensure an empty file on failure. However, we would then be trusting
that a) bamUtil returns error codes for all errors; b) the pipeline is
running in "harsh" mode.