Remove linkArtefacts symlinking functionality for now
Leaving this in Git history for easier restore. Currently not required
and complicates things - using it to fix downstream tools would require
more care with clashes and ensuring that links are updated as the
pipeline runs. bamMetrics is also currently the only annoying
candidate (_dedup in directory name) and can potentially just be fixed
on its own.