Add more tests
This adds a substantial amount of testing to the pipeline itself, and
some static analysis to the jobs:

* "unit" tests of some functionality
* integration tests of complete pipeline runs with:
  * FASTQ vs BAM input
  * every INI file in the settings directory
  * with/without .done files present
  * bash syntax checks on the jobs
  * shellcheck bash linting on the jobs
* possibility of running coverage

Note that coverage is grossly underestimated due to the integration
tests not being counted (they run in a separate process - we could also
run the full pipeline by calling the function, but that would be almost
the same and miss any edge cases around calling the script).

The number of tests reported by the harness is rather high and variable,
because of the parameterisation described above - removing an INI will
decrease the count by several hundred. On the other hand, perlcritic is
running many policies per file and only reports the number of files
tested.

Either way: there are now tests.