Upgrade FREEC to v10.3
This implicitly sets degree=1, which is the new default for non-GC
based normalisation. We want this change anyway, because it reduces
poor fitting due to cubic and quartic terms in corner cases.

Regression is not perfect but looks superficially acceptable. We have
anyway not really been using the CNV output.