Delly parallelises over samples so maximum 2 threads
Most parallelisation is gained by splitting over chromosomes within
the pipeline. Setting the correct number allows the right number of
SGE jobs to run concurrently.

Ideally this is set to 1 for single sample but doesn't seem worth
forking the setting for now (it would actually be much better to add a
qsub function that allows a scale factor to be applied to the base
_THREADS setting and set the base to 1 for Delly).