Add missing PostStats "parent" done file, ensure that markDone depends on all jobs in chain
markDone marks the done file of a chain jobs which form one bigger
job. It is probably not necessary if all samples are re-run with more
fine-grained jobs and thus more fine-grained, renamed done files.

However, when it is used it should depend on all potential jobs in the
chain, not just the last one - otherwise if the last one is skipped
the markDone job will run too early. Other jobs which use the stored
value of markDone's job ID from RUNNING_JOBS will also then run too
early.