Initial support for ignoring decoys
This is a bit of a hack to prevent a very large increase in Delly
jobs (to the point where the command line is too long to submit
combining jobs). It is not complete decoy support.

Currently the pipeline will run fine with decoys. Without this change
it will submit calling jobs for decoy chromosomes, which can also be
fine. The extra run time may even not be an issue in most cases.

Other approaches would be necessary for complete support: e.g. either
splitting off the decoy mappings (and saving them) from the BAM
post-mapping, or filtering final VCFs against this "core" genome (or a
BED file produced from it). This just seemed like the minimal change
to keep things sane. It does not prevent un-split callers from
examining the decoys (e.g. Strelka, Manta).

It's not completely clear that all tools after mapping would work by
simply supplying the core genome in place of the full genome unless
the BAM is split.