Split out pre-calling steps (pileup)
Creating the pileup is not inherently part of somatic calling, it is
only a dependency of one the callers. Pileup may also be used by
FREEC, for example. Any other jobs should be able to depend on it
completing and lookup its output files.

This could also be an appropriate place to add e.g. readcount file
creation.

Must come after all BAM preparation steps.