Add FREEC BAF/pileup option
When FREEC runs from pileup it can produce BAF output. Should also be
quite a bit faster. Add this as a supported option.

The default normalization method in this mode switches to GC-content,
which is recommended by the author for WGS when BAF output is
available and is what we want as well.