Enforce bamMetrics synchronisation in the already-scheduled job
Running this in the lightweight check job doesn't seem necessary and
causes the check to be run with (unnecessarily high) resources
allocated whilst just waiting for bamMetrics. Waiting in the main job
means that the resources used "waiting" are the resources actually
allocated to the computation - effectively we just enforce
synchronisation on bamMetrics.

Standardisation of job creation/submission means the job ID of the
check job is not available at the time of creation of the main job -
this is rather circular dependency that is not worth adding
corner-cases for. Ultimately caused by bamMetrics not waiting for its
own completion - it should really just be part of the pipeline.

bamMetrics. urgh.