Remove aggressive flag from vt decompose_blocksub
The aggressive flag is marked as being designed for "callers like
Freebayes". However, although it doesn't appear to do anything
incorrect, it removes information the extent that vt normalize sees
the normalisation as ambiguous, and converts to alternate
interpretation. Both of these are correct but inconsistent with the
other callers/truth set representation.

This accounts for ~3000 extra calls in the GIAB mix-in, and removing
this flag removes those calls. Checking individual cases shows
that (where correct) they again align with the truth set.