Add the ability to keep BQSR BAM file
Much like the pileup, the BQSR BAM is very large and not worth the
cost of storage. However, it also takes a long time to produce. When
doing experiments - such a re-parameterisations of previous samples or
adding new tools - BQSR is even worse than pileup because the BAM is
essentially needed by all tools, whereas the pileup is only needed by
Varscan (and FREEC if using BAF). Allowing the BAM to be optionally
kept reduces wasted compute in such situations. It can also be
automatically enabled in settings.config by e.g. run_pipeline scripts
in the same way that FINALIZE_KEEP_PILEUP is now. It will then
override the default setting for the INI being run without needing a
separate INI.