Make links.json use relative paths
This means that the links are portable - they do not need to be
changed as the run directory is transferred between
directories/machines. We already had to write code to do this in two
or three different places - using absolute paths was ill-advised. You
always know the parent directory when accessing the links.json, so it
is trivial to derive the full path.